Move JSON response body processing to JS http module
This allows us to use ToBytes to process any body type without adding
the js/common dependency to httpext, and it arguably should've been part
of the http module to begin with.
